Material UI and simpler player hand (#25)
The tactile approach towards modeling the player hand was a tempting idea, but it ended up too difficult to sync state across the network, handle animations triggered by changes in state (e.g., cards leaving the hand), and handling multiple submissions. This PR simplifies the UI drastically, opting for Material UI and simple lists/checkboxes for submissions. I've left the old PlayerHand component in because I'm still too attached to it (I spent weeks trying to get that swipe interaction to be smooth!)